chore: adjust the logic of css-scope
