+++ packages/css-scope/src/animationParser.ts
