+++ packages/css-scope/src/globalTypes.ts
